make the sliders in the config hud smooth
(cherry picked from commit b3e82e11fa7e45a62b472ede0760bde5d541c7e6)

+++ src/main/java/fr/alexdoru/megawallsenhancementsmod/gui/guiscreens/HitboxConfigGuiScreen.java
@@ -141,7 +141,7 @@ public class HitboxConfigGuiScreen extends MyGuiScreen implements GuiSlider.ISli
     @Override
     public void onChangeSliderValue(GuiSlider slider) {
         if (slider.id == 16) {
-            ConfigHandler.hitboxDrawRange = (float) slider.getValue();
+            ConfigHandler.hitboxDrawRange = slider.getValue();
         }
     }
